Split PlayableDisc from Player
This change does the following:
- Splits out all image reading and manipulation into a new file called PlayableDisc
- Fixes some issues present in loading a disc with a data first track
- Fixes some initialization issues that could be present
- Removed the Player code from being able to directly manipulate most of the disc state
